requirements.txt
flask
line-bot-sdk
gunicorn
pytesseract
pillow
requests